FUSETOOLS2-831 - hover on option for API based components
