+++ src/main/java/com/github/cameltooling/lsp/internal/completion/CamelOptionNamesCompletionsFuture.java
@@ -17,12 +17,7 @@
 package com.github.cameltooling.lsp.internal.completion;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -30,18 +25,12 @@ import java.util.stream.Collectors;
 import org.apache.camel.catalog.CamelCatalog;
 import org.eclipse.lsp4j.CompletionItem;
 
-import com.github.cameltooling.lsp.internal.catalog.model.ApiOptionModel;
-import com.github.cameltooling.lsp.internal.catalog.model.ApiPropertyMethodOptionModel;
-import com.github.cameltooling.lsp.internal.catalog.model.ApiPropertyOptionModel;
 import com.github.cameltooling.lsp.internal.catalog.model.BaseOptionModel;
 import com.github.cameltooling.lsp.internal.catalog.model.ComponentModel;
-import com.github.cameltooling.lsp.internal.catalog.model.EndpointOptionModel;
 import com.github.cameltooling.lsp.internal.catalog.util.ModelHelper;
-import com.github.cameltooling.lsp.internal.instancemodel.CamelURIInstance;
 import com.github.cameltooling.lsp.internal.instancemodel.CamelUriElementInstance;
 import com.github.cameltooling.lsp.internal.instancemodel.OptionParamKeyURIInstance;
 import com.github.cameltooling.lsp.internal.instancemodel.OptionParamURIInstance;
-import com.github.cameltooling.lsp.internal.instancemodel.PathParamURIInstance;
 
 public class CamelOptionNamesCompletionsFuture implements Function<CamelCatalog, List<CompletionItem>>  {
 
@@ -98,56 +87,8 @@ public class CamelOptionNamesCompletionsFuture implements Function<CamelCatalog,
 		ComponentModel componentModel = ModelHelper.generateComponentModel(catalog.componentJSonSchema(camelComponentName), true);
 		List<BaseOptionModel> allOptions = new ArrayList<>();
 		allOptions.addAll(componentModel.getEndpointOptions());
-		allOptions.addAll(findAvailableApiProperties(componentModel));
+		allOptions.addAll(uriElement.findAvailableApiProperties(componentModel));
 		return allOptions;
-	}
-
-	private List<EndpointOptionModel> findAvailableApiProperties(ComponentModel componentModel) {
-		CamelURIInstance camelUriInstance = uriElement.getCamelUriInstance();
-		PathParamURIInstance apiNamePath = camelUriInstance.getComponentAndPathUriElementInstance().getApiNamePath();
-		PathParamURIInstance methodNamePath = camelUriInstance.getComponentAndPathUriElementInstance().getMethodNamePath();
-		Optional<ApiPropertyMethodOptionModel> apisPropertiesModel = componentModel.getApiProperties()
-				.stream()
-				.filter(apiProperty -> isCorrespondingApiName(apiNamePath, apiProperty))
-				.map(apiProperty -> findApiPropertyModel(componentModel, methodNamePath, apiProperty))
-				.filter(Objects::nonNull)
-				.findAny();
-		if(apisPropertiesModel.isPresent()) {
-			return apisPropertiesModel.get().getProperties();
-		} else {
-			return Collections.emptyList();
-		}
-	}
-
-	private boolean isCorrespondingApiName(PathParamURIInstance apiNamePath, ApiPropertyOptionModel apiProperty) {
-		return apiNamePath != null && apiProperty.getName().equals(apiNamePath.getValue());
-	}
-
-	private ApiPropertyMethodOptionModel findApiPropertyModel(ComponentModel componentModel, PathParamURIInstance methodNamePath, ApiPropertyOptionModel apiProperty) {
-		Optional<ApiOptionModel> correspondingApi = componentModel.getApis()
-				.stream()
-				.filter(api -> apiProperty.getName().equals(api.getName()))
-				.findAny();
-		if(correspondingApi.isPresent()) {
-			Map<String, String> aliasesMapping = new HashMap<>();
-			for(String aliasFullString : correspondingApi.get().getAliases()) {
-				String[] splittedAlias = aliasFullString.split("=");
-				aliasesMapping.put(splittedAlias[1], splittedAlias[0]);
-			}
-			String methodKind = aliasesMapping.get(methodNamePath.getValue());
-			if("^creator$".equals(methodKind)) {
-				return apiProperty.getCreator();
-			} else if("^deleter$".equals(methodKind)) {
-				return apiProperty.getDeleter();
-			} else if("^fetcher$".equals(methodKind)) {
-				return apiProperty.getFetcher();
-			} else if("^reader$".equals(methodKind)) {
-				return apiProperty.getReader();
-			} else if("^updater$".equals(methodKind)) {
-				return apiProperty.getUpdater();
-			}
-		}
-		return null;
 	}
 
 }
